Something something switch to using Bucklescript instead of js_of_ocaml.
This took longer than expected mostly because I didn't know
bucklescript's implementation of variants couldn't be serialized.
Variants, through BS, become arrays with an added property `tag`. That
is used to figure out what that variant is. The problem is that
`JSON.stringify` doesn't know about that property on arrays, and will
not serialize it.
So `JSON.parse(JSON.stringify(variant)).tag == 0` will always be true.
This took a couple of hours to figure out, and the work around is to
detect when variants are sent over the wire and serialize them properly.

There are tons of fixes that need to be pushed to resocket.io and
reason-js.

Also this only worked after lots of fiddling, but something's wrong
with the BS + Rebel installation process. I had to use `npm link` magic
to get it to compile.

Will fix ... when I wake up.