Add grid. Updating printing strategy. Add new sprite.
I added a grid because we didn't have one. We need the concept of a grid
because we're going to generate dungeons. The simplest possible
representation of a dungeon is a massive 2D grid with tiles saying
what to draw, so I went with that. We'll probably revisit at some point
because this is a very inefficient DS.

I updating printing by currying a print function at the beginning of
any function with the name of the function. This makes a nice print
by default with the name of the function printing before the print.

I added a new sprite for `Empty` cells because I needed to see
something. Very programmer art. Aint got no time.